requirements.txt
flask==3.0.3
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
opencv-python==4.10.0.84
flask-cors==4.0.0